"Added city names to homepage view by querying database in AppController"
This diff adds a new database query in the `homepage` method of the `AppController` class to retrieve all records from the `city` table. The result of this query is then passed to the `homepage` view and displayed in an unordered list. The query is implemented using the `DB` facade's `select` method, which allows for raw SQL queries to be executed against the database. The new data is added to the existing `$data` array, which already contains the title and movie data. The view is not modified in any other way.