"Updated pagination links to conditionally disable 'Anterior' and 'Próxima' buttons based on current page"
The view file for displaying paginated cities has been updated to include conditional statements for the 'Anterior' and 'Próxima' page links. If the current page is the first page, the 'Anterior' link will be displayed as a disabled span. If the current page is the last page, the 'Próxima' link will be displayed as a disabled span. Otherwise, the links will be active and direct to the previous or next page of results. Additionally, the current page number and total number of pages are now displayed at the bottom of the table.