"Added a form to allow selecting the number of cities to show per page in the paginated city list"
This commit adds a form to the paginatedCities.blade.php view that allows the user to select the number of cities they want to show per page. The form contains a dropdown with options for 20, 50, and 100 cities per page, as well as a submit button to apply the selected option. The purpose of this change is to give users more control over the amount of data displayed on each page of the city list.